requirements.txt
python-dotenv==1.0.0
aiogram~=3.0.0b7